renamed types folder to event-types
