+++ src/app/dataview/event-types/colorer.ts
