+++ src/app/dataview/event-types/types-toolbox.component.html
